Internal refactoring of multisession functions (#3935)
Split the Session manipulation functions between the legacy ones using
the embedded storage and the new ones integrated with external storage.
