+++ packages/node/src/multiSession.ts
@@ -19,10 +19,7 @@
 // SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 //
 
-import type {
-  SessionTokenSet,
-  IStorage,
-} from "@inrupt/solid-client-authn-core";
+import type { IStorage } from "@inrupt/solid-client-authn-core";
 import { EVENTS } from "@inrupt/solid-client-authn-core";
 import type ClientAuthentication from "./ClientAuthentication";
 import { getClientAuthenticationWithDependencies } from "./dependencies";
@@ -94,42 +91,6 @@ export async function refreshSession(
       tokenType: sessionInfo.tokenType,
     });
   }
-}
-
-/**
- * Refresh the Access Token and ID Token using the Refresh Token.
- * The tokens may not be expired in order to be refreshed.
- *
- * @param tokenSet the tokens to refresh
- * @returns a new set of tokens
- * @since 2.4.0
- * @example
- * ```
- * const refreshedTokens = await refreshTokens(previousTokenSet);
- * const session = await Session.fromTokens(refreshedTokens, sessionId);
- * ```
- */
-export async function refreshTokens(tokenSet: SessionTokenSet) {
-  const session = await Session.fromTokens(tokenSet);
-  // Replace with Promise.withResolvers when minimal node is 22.
-  let tokenResolve: (tokens: SessionTokenSet) => void;
-  let tokenReject: (reason?: Error) => void = () => {};
-  const tokenPromise = new Promise<SessionTokenSet>((resolve, reject) => {
-    tokenResolve = resolve;
-    tokenReject = reject;
-  });
-  session.events.on("newTokens", (tokens) => {
-    tokenResolve(tokens);
-  });
-  await session.login({
-    oidcIssuer: tokenSet.issuer,
-    clientId: tokenSet.clientId,
-    refreshToken: tokenSet.refreshToken,
-  });
-  if (!session.info.isLoggedIn) {
-    tokenReject(new Error("Could not refresh the session."));
-  }
-  return tokenPromise;
 }
 
 async function internalGetSessionFromStorage(
