+++ packages/node/src/multisession.fromTokens.ts
@@ -1,0 +1,59 @@
+//
+// Copyright Inrupt Inc.
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal in
+// the Software without restriction, including without limitation the rights to use,
+// copy, modify, merge, publish, distribute, sublicense, and/or sell copies of the
+// Software, and to permit persons to whom the Software is furnished to do so,
+// subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED,
+// INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A
+// PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT
+// HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
+// OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE
+// SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+//
+
+import type { SessionTokenSet } from "@inrupt/solid-client-authn-core";
+import { Session } from "./Session";
+
+/**
+ * Refresh the Access Token and ID Token using the Refresh Token.
+ * The tokens may not be expired in order to be refreshed.
+ *
+ * @param tokenSet the tokens to refresh
+ * @returns a new set of tokens
+ * @since 2.4.0
+ * @example
+ * ```
+ * const refreshedTokens = await refreshTokens(previousTokenSet);
+ * const session = await Session.fromTokens(refreshedTokens, sessionId);
+ * ```
+ */
+export async function refreshTokens(tokenSet: SessionTokenSet) {
+  const session = await Session.fromTokens(tokenSet);
+  // Replace with Promise.withResolvers when minimal node is 22.
+  let tokenResolve: (tokens: SessionTokenSet) => void;
+  let tokenReject: (reason?: Error) => void = () => {};
+  const tokenPromise = new Promise<SessionTokenSet>((resolve, reject) => {
+    tokenResolve = resolve;
+    tokenReject = reject;
+  });
+  session.events.on("newTokens", (tokens) => {
+    tokenResolve(tokens);
+  });
+  await session.login({
+    oidcIssuer: tokenSet.issuer,
+    clientId: tokenSet.clientId,
+    refreshToken: tokenSet.refreshToken,
+  });
+  if (!session.info.isLoggedIn) {
+    tokenReject(new Error("Could not refresh the session."));
+  }
+  return tokenPromise;
+}
